pyproject.toml
[project]
name = "bark-lead-discovery-scorer-and-pitch-generator-ai-agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "langchain-groq>=1.1.2",
    "langgraph>=1.1.3",
    "playwright>=1.58.0",
    "psycopg2-binary>=2.9.11",
    "python-dotenv>=1.2.2",
    "sqlalchemy>=2.0.48",
]
